Move data chunk encoding into TsvWriter with preallocated header
Preallocate the 8-byte frame header (flagged_len + table_id + row_count) at
the start of TsvWriter's buffer so the uncompressed path fills it in-place
with zero copies. Compression (zstd) builds a separate buffer as before.

Also narrows row_count from u32 to u16 in the wire format and replaces the
decode_data_chunk 3-tuple return with a DataChunk struct.

+++ crates/protocol/src/messages.rs
@@ -122,7 +122,7 @@ pub enum ServerMessage {
     /// Data chunk with rows in TSV format
     Data {
         table_id: u16,
-        row_count: u32,
+        row_count: u16,
         tsv_data: Vec<u8>,
     },
     /// Table fully transferred
